Cleaned up dependencies between player and inventory
- Use type-safe signal emitter functions throughout the codebase
- Better separation of concerns between items and player and inventory and player
- Items can be picked, dropped and picked up again

+++ scenes/buffs/Buff.cs
@@ -5,13 +5,17 @@ public abstract partial class Buff : Resource
 {
     [Export] public string Name;
     [Export] public float Duration = 1f;
-    [Export] public int TicksPerSecond = 1;
+    [Export] public float TicksPerSecond = 1;
 
     public void ApplyTo(Player player)
     {
         player.ApplyBuff(this);
     }
 
+    /// <summary>
+    /// Callback to override for custom behavior when 
+    /// the buff is applied to a player.
+    /// </summary>
     public virtual void OnTick(Player player)
     { }
 }
